fix(test): update imports after server/utils → server/utils/files move
Files moved in #371 (refactor/files-io-phase4):
- server/utils/image-store.ts → server/utils/files/image-store.ts
- server/utils/markdown-store.ts → server/utils/files/markdown-store.ts
- server/utils/spreadsheet-store.ts → server/utils/files/spreadsheet-store.ts
